fix: use commonjs for scripts
