+++ packages/lozenge/build.js
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const { build } = require('../../scripts/esbuild.js');
+
+build(require('./package.json'));
