--- packages/lozenge/build.mjs
@@ -1,4 +1,0 @@
-import pkg from './package.json';
-import { build } from '../../scripts/esbuild.mjs';
-
-build(pkg);
